pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "msfabricutils"
version = "0.1.8"
description = "A Python library exposes additional functionality to work with Python Notebooks in Microsoft Fabric."
authors = [
    { name = "Jimmy Jensen" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["fabric", "Microsoft", "duckdb", "deltalake"]
license.file = "LICENSE"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.17.1",
    "duckdb>=1.1.3",
    "deltalake>=0.22.0",
    "sqlglot>=25.32.1",
]

[project.urls]
Repository = "https://github.com/mrjsj/msfabricutils"

[project.optional-dependencies]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.47",
    "mkdocs-gen-files==0.5.0",
    "mkdocstrings-python",
]
dev = [
    "pytest>=6.2.5",
    "ruff>=0.8.1",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = [
  "src",
]

